+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [] - Unreleased
 
+* AWS: Create and delete DNS records
+
 ## [1.0.0] - 2018-02-23
 
 Initial support for multi-region clusters on Azure. See [sye-azure/README.md](sye-azure/README.md) for details.